+++ src/configs/deMorgan.ts
@@ -1,6 +1,19 @@
 import { pluginDeMorgan } from '../eslint'
-import type { ConfigDeMorganOptions, TypedConfigItem } from '../types'
+import type { OptionsOverrides, TypedConfigItem } from '../types'
 
+/**
+ * Options type of {@link configDeMorgan}
+ */
+export type ConfigDeMorganOptions = OptionsOverrides
+
+/**
+ * Config for optimize logic
+ *
+ * @see {@link https://github.com/azat-io/eslint-plugin-de-morgan}
+ *
+ * @param options - {@link ConfigDeMorganOptions}
+ * @returns ESLint configs
+ */
 export const configDeMorgan = (
   options: ConfigDeMorganOptions = {},
 ): TypedConfigItem[] => [
