+++ src/types/config.ts
@@ -2,13 +2,14 @@
  * @file Config options
  */
 
-import type { Options as VueBlocksOptions } from 'eslint-processor-vue-blocks'
 import type {
   ConfigAntfuOptions,
   ConfigCommandOptions,
+  ConfigDeMorganOptions,
   ConfigDependOptions,
   ConfigESLintCommentsOptions,
   ConfigESLintPluginOptions,
+  ConfigFormatOptions,
   ConfigGitHubActionOptions,
   ConfigGitIgnoreOptions,
   ConfigIgnoresOptions,
@@ -17,225 +18,30 @@ import type {
   ConfigJsdocOptions,
   ConfigJsoncOptions,
   ConfigMarkdownOptions,
+  ConfigNodeOptions,
+  ConfigNtnyqOptions,
   ConfigPerfectionistOptions,
+  ConfigPiniaOptions,
+  ConfigPrettierOptions,
   ConfigRegexpOptions,
+  ConfigSortOptions,
+  ConfigSpecialsOptions,
   ConfigSVGOOptions,
+  ConfigTestOptions,
   ConfigTomlOptions,
+  ConfigTypeScriptOptions,
   ConfigUnicornOptions,
   ConfigUnoCSSOptions,
+  ConfigUnusedImportsOptions,
+  ConfigVueOptions,
   ConfigYmlOptions,
 } from '../configs'
-import type { PrettierOptions } from '../types'
-import type {
-  ESLintRuleSeverity,
-  TSESLintParserOptions,
-  TypedConfigItem,
-} from './eslint'
-import type {
-  OptionsExtensions,
-  OptionsFeatures,
-  OptionsFiles,
-  OptionsOverrides,
-} from './options'
-
-export interface ConfigDeMorganOptions extends OptionsOverrides {}
-
-export interface ConfigFormatOptions {
-  /**
-   * Enable formatter support for css, less, scss, sass and etc.
-   *
-   * @default true
-   */
-  css?: boolean
-
-  /**
-   * Enable formatter support for html
-   *
-   * @default true
-   */
-  html?: boolean
-
-  /**
-   * Options for prettier
-   */
-  prettierOptions?: PrettierOptions
-}
-
-export interface ConfigNodeOptions extends OptionsOverrides {}
-
-export interface ConfigNtnyqOptions extends OptionsOverrides {}
-
-export interface ConfigPiniaOptions extends OptionsFiles, OptionsOverrides {}
-
-export interface ConfigPrettierOptions extends OptionsOverrides {
-  /**
-   * Glob of built-in disabled files
-   *
-   * @default all svg, toml, svelte and astro files
-   */
-  disabledFiles?: string[]
-
-  /**
-   * rule severity
-   *
-   * @default `warn`
-   */
-  severity?: ESLintRuleSeverity
-
-  /**
-   * Glob of user custom disabled files
-   *
-   * @default []
-   */
-  userDisabledFiles?: string[]
-}
-
-export interface ConfigSortOptions {
-  /**
-   * JSON files to sort all properties alphabeta
-   */
-  additionalJsonFiles?: string[]
-
-  /**
-   * YAML files to sort all properties alphabeta
-   */
-  additionalYamlFiles?: string[]
-
-  /**
-   * @default true
-   */
-  i18nLocale?: boolean
-
-  /**
-   * @default true
-   */
-  packageJson?: boolean
-
-  /**
-   * @default true
-   */
-  pnpmWorkspace?: boolean
-
-  /**
-   * @default true
-   */
-  tsconfig?: boolean
-}
-
-export interface ConfigSpecialsOptions {
-  /**
-   * Overrides cli rules
-   */
-  overridesCliRules?: TypedConfigItem['rules']
-
-  /**
-   * Overrides config files rules
-   */
-  overridesConfigFileRules?: TypedConfigItem['rules']
-
-  /**
-   * Overrides scripts rules
-   */
-  overridesScriptsRules?: TypedConfigItem['rules']
-
-  /**
-   * Overrides user scripts rules
-   */
-  overridesUserScriptsRules?: TypedConfigItem['rules']
-
-  /**
-   * More special case configs
-   */
-  specialCaseConfigs?: TypedConfigItem[]
-
-  /**
-   * ShadcnVue config
-   */
-  shadcnVue?:
-    | boolean
-    | {
-        files?: TypedConfigItem['files']
-        overridesRules?: TypedConfigItem['rules']
-      }
-}
-
-export interface ConfigTestOptions extends OptionsFiles, OptionsOverrides {
-  /**
-   * Overrides built-in vitest rules
-   */
-  overridesVitestRules?: TypedConfigItem['rules']
-
-  /**
-   * enable vitest plugin rules
-   *
-   * @default true if vitest in deps
-   */
-  vitest?: boolean
-}
-
-export interface ConfigTypeScriptOptions
-  extends OptionsExtensions,
-    OptionsFiles,
-    OptionsOverrides {
-  /**
-   * Glob patterns for files that should be type aware.
-   * @default ['**\/*.{ts,tsx}']
-   */
-  filesTypeAware?: string[]
-
-  /**
-   * Glob patterns for files that should not be type aware.
-   * @default ['**\/*.md\/**', '**\/*.astro/*.ts']
-   */
-  ignoresTypeAware?: string[]
-
-  /**
-   * Overrides built-in type aware rules
-   */
-  overridesTypeAwareRules?: TypedConfigItem['rules']
-
-  /**
-   * Additional parser options
-   */
-  parserOptions?: TSESLintParserOptions
-
-  /**
-   * Enable type aware check for TypeScript files
-   */
-  tsconfigPath?: string
-}
-
-export interface ConfigUnusedImportsOptions extends OptionsOverrides {}
-
-export interface ConfigVueOptions
-  extends OptionsFeatures,
-    OptionsFiles,
-    OptionsOverrides {
-  /**
-   * Create virtual files for Vue SFC blocks to enable linting.
-   *
-   * @see https://github.com/antfu/eslint-processor-vue-blocks
-   *
-   * @default true
-   */
-  sfcBlocks?: boolean | VueBlocksOptions
-}
-
-/**
- * Internal configs, not enabled
- */
-interface ConfigOptionsInternal {
-  format?: boolean | ConfigFormatOptions
-  unusedImports?: boolean | ConfigUnusedImportsOptions
-}
+import type { OptionsExtensions } from './options'
 
 /**
  * Config factory options
- * @pg
  */
-export interface ConfigOptions
-  extends ConfigOptionsInternal,
-    OptionsExtensions {
+export interface ConfigOptions extends OptionsExtensions {
   command?: ConfigCommandOptions
   eslintComments?: ConfigESLintCommentsOptions
   ignores?: ConfigIgnoresOptions
@@ -271,4 +77,12 @@ export interface ConfigOptions
   unocss?: boolean | ConfigUnoCSSOptions
   vue?: boolean | ConfigVueOptions
   yml?: boolean | ConfigYmlOptions
+}
+
+/**
+ * Not enabled for now, maybe later
+ */
+export interface ConfigOptionsInternal {
+  format?: boolean | ConfigFormatOptions
+  unusedImports?: boolean | ConfigUnusedImportsOptions
 }
